hide MemoryUsageBasedConnectionLimits
