+++ misc/memory-connection-limits/tests/max_bytes.rs
@@ -34,20 +34,18 @@ fn max_bytes() {
     let max_allowed_bytes = connection_limit * 1024 * 1024;
 
     let mut network = Swarm::new_ephemeral(|_| TestBehaviour {
-        connection_limits: Behaviour::new(
-            MemoryUsageBasedConnectionLimits::default().with_max_bytes(max_allowed_bytes),
-        )
-        .with_memory_usage_refresh_interval(None),
+        connection_limits: Behaviour::new_with_max_bytes(max_allowed_bytes)
+            .with_memory_usage_refresh_interval(None),
         mem: Default::default(),
     });
 
     // Adds current mem usage to the limit and update
     let current_mem = memory_stats::memory_stats().unwrap().physical_mem;
     let max_allowed_bytes_plus_base_usage = max_allowed_bytes + current_mem;
-    network.behaviour_mut().connection_limits.update_limits(
-        MemoryUsageBasedConnectionLimits::default()
-            .with_max_bytes(max_allowed_bytes_plus_base_usage),
-    );
+    network
+        .behaviour_mut()
+        .connection_limits
+        .update_max_bytes(max_allowed_bytes_plus_base_usage);
 
     let addr: Multiaddr = "/memory/1234".parse().unwrap();
     let target = PeerId::random();
